miniconda

+++ .gitpod.Dockerfile
@@ -4,17 +4,23 @@ RUN sudo apt-get update && sudo apt-get install -y zsh && \
     wget https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh -O - | zsh && \
     rm -rf install.sh
 
-RUN sudo apt-get update --fix-missing && \
-    sudo apt-get install -y --no-install-recommends wget bzip2 ca-certificates libglib2.0-0 libxext6 libsm6 libxrender1 git mercurial subversion && \
-    sudo apt-get clean && \
-    sudo rm -rf /var/lib/apt/lists/* && \
-    sudo wget --quiet https://repo.anaconda.com/archive/Anaconda3-2021.05-Linux-x86_64.sh -O ~/anaconda.sh && \
-    sudo /bin/bash ~/anaconda.sh -b -p /opt/conda && \
-    sudo rm ~/anaconda.sh && \
+ENV PATH /opt/conda/bin:$PATH
+
+# Leave these args here to better use the Docker build cache
+ARG CONDA_VERSION=py38_4.9.2
+ARG CONDA_MD5=122c8c9beb51e124ab32a0fa6426c656
+
+RUN sudo wget --quiet https://repo.anaconda.com/miniconda/Miniconda3-${CONDA_VERSION}-Linux-x86_64.sh -O miniconda.sh && \
+    echo "${CONDA_MD5}  miniconda.sh" > miniconda.md5 && \
+    if ! md5sum --status -c miniconda.md5; then exit 1; fi && \
+    sudo mkdir -p /opt && \
+    sudo sh miniconda.sh -b -p /opt/conda && \
+    sudo rm miniconda.sh miniconda.md5 && \
     sudo ln -s /opt/conda/etc/profile.d/conda.sh /etc/profile.d/conda.sh && \
     echo ". /opt/conda/etc/profile.d/conda.sh" >> ~/.bashrc && \
     echo "conda activate base" >> ~/.bashrc && \
     sudo find /opt/conda/ -follow -type f -name '*.a' -delete && \
     sudo find /opt/conda/ -follow -type f -name '*.js.map' -delete && \
     sudo /opt/conda/bin/conda clean -afy && \
+    sudo wget --quiet https://public-bucket-all-purpose.s3.fr-par.scw.cloud/data/creditcard.csv resources/data/creditcard.csv && \
     pip install mlflow
